♻️ [Refactoring]: group plugin/ meta value objects under plugin/meta/
Move meta.rs / manifest.rs / manifest_resolve.rs / version.rs into plugin/meta/.
Introduce plugin/meta.rs as the sub-parent. Declare the inner `meta` module as
private (`mod meta;`) and lift the leaf `pub` symbols via `pub use self::meta::*;`
so `crate::plugin::meta::load_meta` etc. keep pointing to the leaf functions
(ref 論点 b). Declare `version` as `pub(crate) mod version;` so the parent
`plugin.rs`'s `pub(crate) use meta::version;` compiles (rustc rejects
re-exporting a private module with E0365); this is crate-internal only and adds
no external `pub mod version;` path (ref 論点 c). The parent re-exports
`{Author, PluginManifest, fetch_remote_versions, PluginMeta, UpgradeState}`
through `meta` so existing external paths `crate::plugin::Author` etc. remain
unchanged. `update.rs:9`'s `use crate::plugin::version::{...}` is left intact
thanks to `pub(crate) use meta::version;` on plugin.rs. Apply
#[allow(clippy::module_inception)] to the same-name leaf module. Rewrite the
temporary `crate::plugin::manifest_resolve::*` references in cache/cache.rs
(introduced in commit 1) to `crate::plugin::meta::*`.

Refs #246
