+++ src/plugin/meta/manifest.rs
