+++ src/plugin/meta/manifest_resolve.rs
