+++ src/plugin/meta/manifest_resolve_test.rs
