+++ src/plugin/meta/manifest_test.rs
